Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 void main(){
-    float x;
-    scanf("%f",&x);
+    double x;
+    scanf("%lf",&x);
     printf("You entered: %.15lf",x);
     return 0;
 }
